https://guidewirejira.atlassian.net/browse/ISGOSU-508 - move StringPool to internal package, tag with @UnstableAPI - reenable caching
