+++ gosu-core-api/src/main/java/gw/lang/parser/exceptions/ParseWarningForDeprecatedMember.java
@@ -6,7 +6,7 @@ package gw.lang.parser.exceptions;
 
 import gw.lang.parser.resources.Res;
 import gw.lang.parser.IParserState;
-import gw.util.cache.StringPool;
+import gw.internal.gosu.util.StringPool;
 
 public class ParseWarningForDeprecatedMember extends ParseWarning
 {
